Make it testable
Both via STDIO and a test harness

Also includes some drive-by changes like unbreaking `start`, better
plural phrasing, and better message combining.
